Fixed errors.

+++ project2-back/src/main/java/net/revature/project1/controller/PostController.java
@@ -90,10 +90,10 @@ public class PostController {
     }
 
     @GetMapping("{id}/comments/total")
-    public ResponseEntity<Long> returnTotalComments(@PathVariable Long id){
-        Long result = postService.returnTotalComments(id);
+    public ResponseEntity<Integer> returnTotalComments(@PathVariable Long id){
+        Integer result = postService.returnTotalComments(id);
         if(result == null){
-            return ResponseEntity.badRequest().body((long) -1);
+            return ResponseEntity.badRequest().body(-1);
         }
         return ResponseEntity.ok(result);
     }
